src/celery_service/pyproject.toml
[tool.poetry]
name = "alive-check-celery"
version = "0.1.0"
description = "Service for retrieve and send ip addresses to ping"
authors = ["Jakub Jadczak <jakubjadczak02@aoutlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
pika = "^1.3.2"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
